not working

+++ airflow/plugins/operators/stage_redshift.py
@@ -2,6 +2,7 @@ from airflow.contrib.hooks.aws_hook import AwsHook
 from airflow.hooks.postgres_hook import PostgresHook
 from airflow.models import BaseOperator
 from airflow.utils.decorators import apply_defaults
+from datetime import datetime
 
 
 class StageToRedshiftOperator(BaseOperator):
@@ -22,6 +23,7 @@ class StageToRedshiftOperator(BaseOperator):
                  s3_bucket="",
                  s3_key="",
                  json="auto",
+                 run_once=False,
                  *args, **kwargs):
 
         super(StageToRedshiftOperator, self).__init__(*args, **kwargs)
@@ -31,31 +33,39 @@ class StageToRedshiftOperator(BaseOperator):
         self.s3_bucket = s3_bucket
         self.s3_key = s3_key
         self.json = json
+        self.run_once = run_once
 
     def execute(self, context):
+
         self.log.info('StageToRedshiftOperator starting...')
-        aws_hook = AwsHook(self.aws_credentials_id)
-        credentials = aws_hook.get_credentials()
-        redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
 
-        self.log.info("Clearing data from " +
-                      f"Redshift table: {self.table}")
-        redshift.run(f"DELETE FROM {self.table}")
+        prev_success = context.get('prev_ds')
+        self.log.info("Previous Completion Interval Start: " + str(prev_success))
+        if self.run_once and prev_success:
+            self.log.info("Task Previously completed. Skipping.")
+        else:
 
-        self.log.info("Copying data from S3 to Redshift")
-        # need to put in year/month/hour (or something)
-        execution_date = context['ds']
-        rendered_key = self.s3_key.format(execution_date.year,
-                                          execution_date.month,
-                                          execution_date.year,
-                                          execution_date.month,
-                                          execution_date.day)
-        s3_path = "s3://{}/{}".format(self.s3_bucket, rendered_key)
-        formatted_sql = StageToRedshiftOperator.copy_sql.format(
-            self.table,
-            s3_path,
-            credentials.access_key,
-            credentials.secret_key,
-            self.json
-        )
-        redshift.run(formatted_sql)
+            aws_hook = AwsHook(self.aws_credentials_id)
+            credentials = aws_hook.get_credentials()
+            redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
+
+            self.log.info(f"Clearing data from table: {self.table}")
+            redshift.run(f"DELETE FROM {self.table}")
+
+            self.log.info("Copying data from S3 to Redshift")
+
+            execution_date = datetime.strptime(context['ds'], '%Y-%m-%d')
+            rendered_key = self.s3_key.format(execution_date.year,
+                                              execution_date.month,
+                                              execution_date.year,
+                                              execution_date.month,
+                                              execution_date.strftime("%d"))
+            s3_path = "s3://{}/{}".format(self.s3_bucket, rendered_key)
+            formatted_sql = self.copy_sql.format(
+                self.table,
+                s3_path,
+                credentials.access_key,
+                credentials.secret_key,
+                self.json
+            )
+            redshift.run(formatted_sql)
